Implement user CRUD with validation and pagination

+++ src/main/java/com/example/service/AuthorService.java
@@ -2,10 +2,13 @@ package com.example.service;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import com.example.dto.request.AuthorRequestDto;
 import com.example.dto.response.AuthorResponseDto;
+import com.example.exception.DuplicateResourcesException;
 import com.example.exception.ResourcesNotFoundException;
 import com.example.model.Author;
 import com.example.repository.AuthorRepository;
@@ -23,6 +26,13 @@ public class AuthorService {
     }
 
     public AuthorResponseDto save(AuthorRequestDto request){
+
+        if(authorRepository.existsByPhoneNumber(request.getPhoneNumber())){
+               throw new DuplicateResourcesException("Phone Number Already Exists " +request.getPhoneNumber());
+        }
+        if(authorRepository.existsByEmail(request.getEmail())){
+            throw new DuplicateResourcesException("Author  Email Already Exists " +request.getEmail());
+        }
 
         Author author = new Author();
         {
@@ -48,6 +58,12 @@ public class AuthorService {
     public List<AuthorResponseDto> saveAll(List<AuthorRequestDto> requests){
         List<Author> authors = requests.stream()
             .map(request -> {
+                 if(authorRepository.existsByPhoneNumber(request.getPhoneNumber())){
+                        throw new DuplicateResourcesException("Phone Number Already Exists " +request.getPhoneNumber());
+                 }
+                 if(authorRepository.existsByEmail(request.getEmail())){
+                        throw new DuplicateResourcesException("Author  Email Already Exists " +request.getEmail());
+                 }
                 Author author = new Author();
 
                 author.setName(request.getName());
@@ -104,20 +120,23 @@ public class AuthorService {
             author.getNationality()
         );
     }
-    public List<AuthorResponseDto> findall(){
-        List<Author> authors = authorRepository.findAll();
+    public Page<AuthorResponseDto> findall(Pageable pageable){
+        Page<Author> authors = authorRepository.findAll(pageable);
         if(authors.isEmpty()){
             throw new ResourcesNotFoundException("Authors Not found");
         }
-        return authors.stream().map(author -> new AuthorResponseDto(
+        return authors.map(author -> new AuthorResponseDto(
             author.getAuthorId(),
             author.getName(),
             author.getEmail(),
             author.getNationality()
-        )).collect(Collectors.toList());
+        ));
     }
 
     public void deletdeById(String id){
+        if(!authorRepository.existsById(id)){
+            throw new ResourcesNotFoundException("Author with Id Not Found" +id); 
+        }
         authorRepository.deleteById(id);
     }
 }
